refactor(editor): move the editor's pure helpers and chrome out of EditorScreen
