+++ app/src/main/java/com/markleaf/notes/feature/editor/EditorDialogs.kt
@@ -1,0 +1,89 @@
+package com.markleaf.notes.feature.editor
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.markleaf.notes.R
+
+/**
+ * The image alt-text editor, opened by a long press on a preview image. Holds
+ * its own draft so typing is not committed until the user confirms.
+ */
+@Composable
+internal fun ImageAltTextDialog(
+    path: String,
+    currentAlt: String,
+    onConfirm: (String) -> Unit,
+    onDismiss: () -> Unit
+) {
+    var draft by remember(path) { mutableStateOf(currentAlt) }
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = { Text(stringResource(R.string.image_alt_dialog_title)) },
+        text = {
+            Column {
+                Text(
+                    text = stringResource(R.string.image_alt_dialog_description),
+                    style = MaterialTheme.typography.bodySmall,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+                Spacer(Modifier.height(12.dp))
+                OutlinedTextField(
+                    value = draft,
+                    onValueChange = { draft = it },
+                    label = { Text(stringResource(R.string.image_alt_dialog_label)) },
+                    singleLine = true,
+                    modifier = Modifier.fillMaxWidth()
+                )
+            }
+        },
+        confirmButton = {
+            Button(onClick = { onConfirm(draft) }) {
+                Text(stringResource(R.string.image_alt_dialog_save))
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = onDismiss) {
+                Text(stringResource(R.string.cancel))
+            }
+        }
+    )
+}
+
+/** The move-to-trash confirmation shown from the editor's overflow menu. */
+@Composable
+internal fun DeleteConfirmDialog(
+    onConfirm: () -> Unit,
+    onDismiss: () -> Unit
+) {
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = { Text(stringResource(R.string.move_to_trash_title)) },
+        text = { Text(stringResource(R.string.move_to_trash_editor_message)) },
+        confirmButton = {
+            Button(onClick = onConfirm) {
+                Text(stringResource(R.string.move_to_trash))
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = onDismiss) {
+                Text(stringResource(R.string.cancel))
+            }
+        }
+    )
+}
